requirements.txt
Django~=2.0.13
djangorestframework==3.10.0
requests==2.22.0
pytest-django==3.5.1
  pytest==5.0.1
django-categories==1.6.1
python-dateutil==2.8.0